feat(db): implement session management logic and connection state pattern
Use Argon2 for password hashing, and BLAKE3 for session token hashing.
Knowing that passwords are likely to have low entropy and session tokens
have high entropy, this should provide the optimal design for both
security and performance. Moreover, `rayon` is utilized to provide a
thread pool to handle hashing calculations without blocking the runtime
with CPU bound tasks. Additionally, `base64ct` is used to encode session
tokens rather than normal `base64` for secure constant-time operation.

Use `secrecy` to ensure zeroing memory after dropping secrets. However, I
don't think that this can be maintained throw the application (due to the
limitations when interacting with the HTTP server and Actix), so it is
not end-to-end protection.

Database logic security is managed by the type system. Depending on the
authenticated user, the connection will have different state, which will
prevent accessing unprivileged methods with compile-time checks. Also,
during session loading, the API will get a `LoggedIn` enum, which will
give the ability to handle each case in a clear and auditable manner.