feat(db/migration): add `ToChangePassword` column to `User` table
This will be used to ask newly created users to change their temporary
backend generated password, and for password resets done by the admin.

Note that `Applicant` users are the only ones who create their initial
passwords by themselves. `RecruitmentManager` users are created by the
admin and `Interviewer` users are created by a `RecruitmentManager`, both
randomly generate user identifier and password to be sent to the user for
initial login.